Move testutils to separate file
